+++ src/reader/storage.py
@@ -79,7 +79,9 @@ class Storage:
         cursor = self.db.execute("""
             SELECT url, updated, title, link, author, user_title FROM feeds
             {where_url_snippet}
-            ORDER BY feeds.title, feeds.url;
+            ORDER BY
+                coalesce(feeds.user_title, feeds.title),
+                feeds.url;
         """.format(**locals()), locals())
 
         for row in cursor: